// src/model/system/orbits/tags/types.ts
export type OrbitTag =
  | 'frontier'
  // common resources
  | 'farming'
  | 'mining'
  | 'industrial'
  | 'trade'
  // rare resources
  | 'exotic gas'
  | 'rare crystals'
  | 'volatile motes'
  | 'zro'
  // rare services
  | 'research'
  | 'cultural'
  | 'shipyards'
  // special orbits
  | 'fortress'
  | 'prison'
  | 'paradise'
  | 'failed outpost'
  | 'ancient ruins'
  | 'pilgrimage'
  // hazards
  | 'storms'
  | 'meteors'
  | 'raiders'
  | 'plague'
  | 'rebellion'
  | 'battleground'
  | 'wildlife'
  | 'corruption'
  | 'criminals'
  | 'doomed'
  // citizens
  | 'primitives'
  | 'refugees'